requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asyncpg==0.29.0
click==8.1.7
fastapi==0.114.2
greenlet==3.1.0
h11==0.14.0
idna==3.9
pydantic==2.9.1
pydantic_core==2.23.3
python-dotenv==1.0.1
python-multipart==0.0.10
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6
